+++ src/api/hunt/getAllTargetProperties.ts
